+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add close() in update_status to force use of a fresh connection on the next update
 - chmod's the output files using concurrent.futures
 - get the Pagurus filewriter tests working again, also disentangle filewriter class from script.
+- return both outfiles and complete failed tasks' folder to user
 
 ## 1.1.2 Summary [05/15/2023]